reformatted code to account for new user schema requirements
