+++ src/utils/consts.js
@@ -162,6 +162,17 @@ const dogSchema = z.object({
 });
 
 /**
+ * Zod object for validating request bodies for users
+ */
+const userSchema = z.object({
+  name: z.string().optional(),
+  email: z.string().optional(),
+  image: z.string().optional(),
+  emailVerified: z.boolean().default(null).optional(),
+  role: z.array(z.enum(consts.roleArray)).optional(),
+});
+
+/**
  * Mock data for testing without any backend interaction
  */
 const mocks = {
@@ -239,4 +250,4 @@ const mocks = {
   ],
 };
 
-export { pages, consts, dogSchema, mocks };
+export { pages, consts, dogSchema, userSchema, mocks };
